🎨 style: 파일, 폴더 분리 및 코드 구조 단순화 #3
